+++ nornir_srl/connections/srlinux.py
@@ -2,10 +2,7 @@ from typing import TYPE_CHECKING, Any, List, Dict, Optional, Union
 import difflib
 import json
 import re
-import copy
-import datetime
 
-from natsort import natsorted
 import jmespath
 
 from pygnmi.client import gNMIclient
@@ -16,6 +13,7 @@ from nornir.core.exceptions import ConnectionException
 
 from .helpers import strip_modules, normalize_gnmi_resp, lpm
 from .network_instance import NetworkInstanceMixin
+from .bgp import BgpMixin
 
 CONNECTION_NAME = "srlinux"
 
@@ -64,7 +62,7 @@ class GnmiPath:
         return None
 
 
-class SrLinux(NetworkInstanceMixin):
+class SrLinux(NetworkInstanceMixin, BgpMixin):
     def open(
         self,
         hostname: Optional[str],
@@ -149,384 +147,6 @@ class SrLinux(NetworkInstanceMixin):
         )
         res = jmespath.search(path_spec["jmespath"], resp[0])
         return {"subinterface": res}
-
-    def get_bgp_rib(
-        self,
-        route_fam: str,
-        route_type: Optional[str] = "2",
-        network_instance: str = "*",
-    ) -> Dict[str, Any]:
-        BGP_RIB_MOD = "bgp-rib"
-        BGP_RIB_MOD2 = "urn:nokia.com:srlinux:bgp:rib-bgp"
-        if self.capabilities is not None:
-            mod_version = [
-                m
-                for m in self.capabilities.get("supported_models", [])
-                if BGP_RIB_MOD in m.get("name") or BGP_RIB_MOD2 in m.get("name")
-            ][0].get("version")
-        else:
-            raise Exception("Cannot get gNMI capabilities")
-
-        BGP_EVPN_VERSION_MAP = {
-            1: ("2021-", "2022-", "2023-", "2024-03", "2024-07"),
-            2: ("20"),
-        }
-        BGP_EVPN_ROUTE_TYPE_MAP = {
-            1: ("2021-", "2022-", "2023-", "2024-03", "2024-07"),
-            2: ("20"),
-        }
-        BGP_IP_VERSION_MAP = {
-            1: ("2021-", "2022-"),
-            2: ("2023-03",),
-            3: ("20"),
-        }
-        ROUTE_FAMILY = {
-            "evpn": "evpn",
-            "ipv4": "ipv4-unicast",
-            "ipv6": "ipv6-unicast",
-        }
-        ROUTE_TYPE_VERSIONS = {
-            1: {
-                "1": "ethernet-ad-routes",
-                "2": "mac-ip-routes",
-                "3": "imet-routes",
-                "4": "ethernet-segment-routes",
-                "5": "ip-prefix-routes",
-            },
-            2: {
-                "1": "ethernet-ad-route",
-                "2": "mac-ip-route",
-                "3": "imet-route",
-                "4": "ethernet-segment-route",
-                "5": "ip-prefix-route",
-            },
-        }
-
-        def augment_routes(d, attribs):  # augment routes with attributes
-            if isinstance(d, list):
-                return [augment_routes(x, attribs) for x in d]
-            elif isinstance(d, dict):
-                if "attr-id" in d:
-                    d.update(attribs.get(d["attr-id"], {}))
-                    d["_r_state"] = (
-                        ("u" if d["used-route"] else "")
-                        + ("*" if d["valid-route"] else "")
-                        + (">" if d["best-route"] else "")
-                    )
-                    if "label1" in d:  # from SRL 24.3 onwards for mac/ip routes
-                        d["vni"] = d["label1"].get("value", "-")
-                    elif "label" in d:  # for SRL 24.3 onwards
-                        d["vni"] = d["label"].get("value", "-")
-                    else:
-                        d["vni"] = d.get("vni", "-")
-                    d["_rt"] = ",".join(
-                        [
-                            x_comm.split("target:")[1]
-                            for x_comm in d.get("communities", {}).get(
-                                "ext-community", []
-                            )
-                            if "target:" in x_comm
-                        ]
-                    )
-                    d["_as_path"] = str(
-                        attribs[d["attr-id"]]
-                        .get("as-path", {})
-                        .get("segment", [{}])[0]
-                        .get("member", [])
-                    )
-                    d["_esi_lbl"] = ",".join(
-                        [
-                            str(x_comm.split("esi-label:")[1])
-                            .replace("Single-Active", "S-A")
-                            .replace("All-Active", "A-A")
-                            for x_comm in d.get("communities", {}).get(
-                                "ext-community", []
-                            )
-                            if "esi-label:" in x_comm
-                        ]
-                    )
-                    return d
-                else:
-                    return {k: augment_routes(v, attribs) for k, v in d.items()}
-            else:
-                return d
-
-        evpn_path_version = [
-            k
-            for k, v in sorted(BGP_EVPN_VERSION_MAP.items(), key=lambda item: item[0])
-            if len([ver for ver in v if mod_version.startswith(ver)]) > 0
-        ][0]
-        evpn_route_type_version = [
-            k
-            for k, v in sorted(
-                BGP_EVPN_ROUTE_TYPE_MAP.items(), key=lambda item: item[0]
-            )
-            if len([ver for ver in v if mod_version.startswith(ver)]) > 0
-        ][0]
-        ip_path_version = [
-            k
-            for k, v in sorted(BGP_IP_VERSION_MAP.items(), key=lambda item: item[0])
-            if len([ver for ver in v if mod_version.startswith(ver)]) > 0
-        ][0]
-
-        if route_fam not in ROUTE_FAMILY:
-            raise ValueError(f"Invalid route family {route_fam}")
-        if (
-            route_type
-            and route_type not in ROUTE_TYPE_VERSIONS[evpn_route_type_version]
-        ):
-            raise ValueError(f"Invalid route type {route_type}")
-
-        PATH_BGP_PATH_ATTRIBS = (
-            "/network-instance[name="
-            + network_instance
-            + "]/bgp-rib/attr-sets/attr-set"
-        )
-        RIB_EVPN_PATH_VERSIONS: Dict[int, Dict[str, Any]] = {
-            1: {
-                "RIB_EVPN_PATH": (
-                    "/network-instance[name=" + network_instance + "]/bgp-rib/"  # type: ignore
-                    f"{ROUTE_FAMILY[route_fam]}/rib-in-out/rib-in-post/"
-                    f"{ROUTE_TYPE_VERSIONS[evpn_route_type_version][route_type]}"  # type: ignore
-                ),
-                "RIB_EVPN_JMESPATH_COMMON": '"network-instance"[].{NI:name, Rib:"bgp-rib"."'
-                + ROUTE_FAMILY[route_fam]
-                + '"."rib-in-out"."rib-in-post"."'
-                + ROUTE_TYPE_VERSIONS[evpn_route_type_version][route_type]  # type: ignore
-                + '"[]',
-                "RIB_EVPN_JMESPATH_ATTRS": {
-                    "1": '.{RD:"route-distinguisher", peer:neighbor, ESI:esi, Tag:"ethernet-tag-id",vni:vni, "NextHop":"next-hop", RT:"_rt", "esi-lbl":"_esi_lbl", "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "2": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, ESI:esi, "MAC":"mac-address", "IP":"ip-address",vni:vni,"next-hop":"next-hop", "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "3": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, Tag:"ethernet-tag-id", "next-hop":"next-hop", origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "4": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, ESI:esi, "next-hop":"next-hop", origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "5": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, lpref:"local-pref", "IP-Pfx":"ip-prefix",vni:vni, med:med, "next-hop":"next-hop", GW:"gateway-ip",origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                },
-            },
-            2: {
-                "RIB_EVPN_PATH": (
-                    "/network-instance[name=" + network_instance + f"]/bgp-rib/afi-safi[afi-safi-name={ROUTE_FAMILY[route_fam]}]/"  # type: ignore
-                    f"{ROUTE_FAMILY[route_fam]}/rib-in-out/rib-in-post/"
-                    f"{ROUTE_TYPE_VERSIONS[evpn_route_type_version][route_type]}"  # type: ignore
-                ),
-                "RIB_EVPN_JMESPATH_COMMON": '"network-instance"[].{NI:name, Rib:"bgp-rib"."afi-safi"[]."'
-                + ROUTE_FAMILY[route_fam]
-                + '"."rib-in-out"."rib-in-post"."'
-                + ROUTE_TYPE_VERSIONS[evpn_route_type_version][route_type]  # type: ignore
-                + '"[]',
-                "RIB_EVPN_JMESPATH_ATTRS": {
-                    "1": '.{RD:"route-distinguisher", peer:neighbor, ESI:esi, Tag:"ethernet-tag-id",vni:vni, "NextHop":"next-hop", RT:"_rt", "esi-lbl":"_esi_lbl", "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "2": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, ESI:esi, "MAC":"mac-address", "IP":"ip-address",vni:vni,"next-hop":"next-hop", "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "3": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, Tag:"ethernet-tag-id", "next-hop":"next-hop", origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "4": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, ESI:esi, "next-hop":"next-hop", origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                    "5": '.{RD:"route-distinguisher", RT:"_rt", peer:neighbor, lpref:"local-pref", "IP-Pfx":"ip-prefix",vni:vni, med:med, "next-hop":"next-hop", GW:"gateway-ip",origin:origin, "0_st":"_r_state", "as-path":"as-path".segment[0].member}}',
-                },
-            },
-        }
-        RIB_IP_PATH_VERSIONS = {
-            1: {
-                "RIB_IP_PATH": (
-                    f"/network-instance[name={network_instance}]/bgp-rib/"
-                    f"{ROUTE_FAMILY[route_fam]}/local-rib/routes"
-                ),
-                "RIB_IP_JMESPATH": '"network-instance"[].{NI:name, Rib:"bgp-rib"."'
-                + ROUTE_FAMILY[route_fam]
-                + '"."local-rib"."routes"[]'
-                + '.{neighbor:neighbor, "0_st":"_r_state", "Prefix":prefix, "lpref":"local-pref", med:med, "next-hop":"next-hop","as-path":"as-path".segment[0].member}}',
-            },
-            2: {
-                "RIB_IP_PATH": (
-                    f"/network-instance[name={network_instance}]/bgp-rib/afi-safi[afi-safi-name={ROUTE_FAMILY[route_fam]}]/"
-                    f"{ROUTE_FAMILY[route_fam]}/local-rib/routes"
-                ),
-                "RIB_IP_JMESPATH": '"network-instance"[].{NI:name, Rib:"bgp-rib"."afi-safi"[]."'
-                + ROUTE_FAMILY[route_fam]
-                + '"."local-rib"."routes"[]'
-                + '.{neighbor:neighbor, "0_st":"_r_state", "Prefix":prefix, "lpref":"local-pref", med:med, "next-hop":"next-hop","as-path":"as-path".segment[0].member,\
-                      "communities":[communities.community, communities."large-community"][]|join(\', \',@)}}',
-            },
-            3: {
-                "RIB_IP_PATH": (
-                    f"/network-instance[name={network_instance}]/bgp-rib/afi-safi[afi-safi-name={ROUTE_FAMILY[route_fam]}]/"
-                    f"{ROUTE_FAMILY[route_fam]}/local-rib/route"
-                ),
-                "RIB_IP_JMESPATH": '"network-instance"[].{NI:name, Rib:"bgp-rib"."afi-safi"[]."'
-                + ROUTE_FAMILY[route_fam]
-                + '"."local-rib"."route"[]'
-                + '.{neighbor:neighbor, "0_st":"_r_state", "Prefix":prefix, "lpref":"local-pref", med:med, "next-hop":"next-hop","as-path":"as-path".segment[0].member,\
-                      "communities":[communities.community, communities."large-community"][]|join(\',\',@)}}',
-            },
-        }
-
-        PATH_SPECS = {
-            "evpn": {
-                "path": RIB_EVPN_PATH_VERSIONS[evpn_path_version]["RIB_EVPN_PATH"],
-                "jmespath": RIB_EVPN_PATH_VERSIONS[evpn_path_version][
-                    "RIB_EVPN_JMESPATH_COMMON"
-                ]
-                + RIB_EVPN_PATH_VERSIONS[evpn_path_version]["RIB_EVPN_JMESPATH_ATTRS"][
-                    route_type
-                ],
-                "datatype": "state",
-            },
-            "ipv4": {
-                "path": RIB_IP_PATH_VERSIONS[ip_path_version]["RIB_IP_PATH"],
-                "jmespath": RIB_IP_PATH_VERSIONS[ip_path_version]["RIB_IP_JMESPATH"],
-                "datatype": "state",
-            },
-            "ipv6": {
-                "path": RIB_IP_PATH_VERSIONS[ip_path_version]["RIB_IP_PATH"],
-                "jmespath": RIB_IP_PATH_VERSIONS[ip_path_version]["RIB_IP_JMESPATH"],
-                "datatype": "state",
-            },
-        }
-
-        attribs: Dict[str, Dict[str, Any]] = dict()
-
-        resp = self.get(paths=[PATH_BGP_PATH_ATTRIBS], datatype="state")
-        for ni in resp[0].get("network-instance", []):
-            if ni["name"] not in attribs:
-                attribs[ni["name"]] = dict()
-            for path in ni.get("bgp-rib", {}).get("attr-sets", {}).get("attr-set", []):
-                path_copy = copy.deepcopy(path)
-                attribs[ni["name"]].update({path_copy.pop("index"): path_copy})
-
-        path_spec: Dict[str, str] = PATH_SPECS[route_fam]
-        resp = self.get(
-            paths=[str(path_spec.get("path"))], datatype=path_spec["datatype"]
-        )
-        for ni in resp[0].get("network-instance", []):
-            ni = augment_routes(ni, attribs[ni["name"]])
-
-        res = jmespath.search(path_spec["jmespath"], resp[0])
-        if res is None:
-            res = []
-        return {"bgp_rib": res}
-
-    def get_sum_bgp(self, network_instance: Optional[str] = "*") -> Dict[str, Any]:
-        BGP_MOD = "urn:srl_nokia/bgp:srl_nokia-bgp"
-        BGP_MOD2 = "urn:nokia.com:srlinux:bgp:bgp:srl_nokia-bgp"
-
-        if self.capabilities is not None:
-            mod_version = [
-                m
-                for m in self.capabilities.get("supported_models", [])
-                if BGP_MOD == m.get("name") or BGP_MOD2 == m.get("name")
-            ][0].get("version")
-        else:
-            raise Exception("Capabilities not set")
-        BGP_VERSION_MAP = {1: ("2021-", "2022-"), 2: ("2023-3", "20")}
-        our_version = [
-            k
-            for k, v in sorted(BGP_VERSION_MAP.items(), key=lambda item: item[0])
-            if len([ver for ver in v if mod_version.startswith(ver)]) > 0
-        ][0]
-
-        def augment_resp(resp):
-            for ni in resp[0].get("network-instance", []):
-                if ni.get("protocols") and ni["protocols"].get("bgp"):
-                    for peer in ni["protocols"]["bgp"]["neighbor"]:
-                        peer_data = dict()
-                        if our_version == 1:
-                            peer_data["evpn"] = peer.get("evpn")
-                            peer_data["ipv4-unicast"] = peer.get("ipv4-unicast")
-                            peer_data["local-as"] = peer.get("local-as", [{}])[0].get(
-                                "as-number", "-"
-                            )
-                        elif our_version == 2:
-                            peer_data["local-as"] = peer.get("local-as", {}).get(
-                                "as-number", "-"
-                            )
-                            for afi in peer.get("afi-safi", []):
-                                if afi["afi-safi-name"] == "evpn":
-                                    peer_data["evpn"] = afi
-                                elif afi["afi-safi-name"] == "ipv4-unicast":
-                                    peer_data["ipv4-unicast"] = afi
-                                elif afi["afi-safi-name"] == "ipv6-unicast":
-                                    peer_data["ipv6-unicast"] = afi
-                        peer["_local-asn"] = peer_data["local-as"]
-                        peer["_flags"] = ""
-                        peer["_flags"] += (
-                            "D" if peer.get("dynamic-neighbor", False) else "-"
-                        )
-                        peer["_flags"] += (
-                            "B"
-                            if peer.get("failure-detection", {}).get(
-                                "enable-bfd", False
-                            )
-                            else "-"
-                        )
-                        peer["_flags"] += (
-                            "F"
-                            if peer.get("failure-detection", {}).get(
-                                "fast-failover", False
-                            )
-                            else "-"
-                        )
-                        if peer_data.get("evpn"):
-                            peer["_evpn"] = (
-                                str(peer_data["evpn"]["received-routes"])
-                                + "/"
-                                + str(peer_data["evpn"]["active-routes"])
-                                + "/"
-                                + str(peer_data["evpn"]["sent-routes"])
-                                if peer_data["evpn"]["admin-state"] == "enable"
-                                else "disabled"
-                            )
-                        else:
-                            peer["_evpn"] = "-"
-                        if peer_data.get("ipv4-unicast"):
-                            if peer_data["ipv4-unicast"]["admin-state"] == "enable":
-                                peer["_ipv4"] = (
-                                    str(peer_data["ipv4-unicast"]["received-routes"])
-                                    + "/"
-                                    + str(peer_data["ipv4-unicast"]["active-routes"])
-                                    + "/"
-                                    + str(peer_data["ipv4-unicast"]["sent-routes"])
-                                )
-                                if (
-                                    peer_data["ipv4-unicast"].get("oper-state")
-                                    == "down"
-                                ):
-                                    peer["_ipv4"] = "down"
-                            else:
-                                peer["_ipv4"] = "disabled"
-                        else:
-                            peer["_ipv4"] = "-"
-                        if peer_data.get("ipv6-unicast"):
-                            if peer_data["ipv6-unicast"]["admin-state"] == "enable":
-                                peer["_ipv6"] = (
-                                    str(peer_data["ipv6-unicast"]["received-routes"])
-                                    + "/"
-                                    + str(peer_data["ipv6-unicast"]["active-routes"])
-                                    + "/"
-                                    + str(peer_data["ipv6-unicast"]["sent-routes"])
-                                )
-                                if (
-                                    peer_data["ipv6-unicast"].get("oper-state")
-                                    == "down"
-                                ):
-                                    peer["_ipv6"] = "down"
-                            else:
-                                peer["_ipv6"] = "disabled"
-                        else:
-                            peer["_ipv6"] = "-"
-
-        path_spec = {
-            "path": f"/network-instance[name={network_instance}]/protocols/bgp/neighbor",
-            "jmespath": '"network-instance"[].{NI:name, Neighbors: protocols.bgp.neighbor[].{"1_peer":"peer-address",\
-                    "peer-as":"peer-as", state:"session-state","local-as":"_local-asn",flags:"_flags",\
-                    "group":"peer-group", "export-policy":"export-policy", "import-policy":"import-policy",\
-                    "AF: IPv4\\nRx/Act/Tx":"_ipv4", "AF: IPv6\\nRx/Act/Tx":"_ipv6", \
-                    "AF: EVPN\\nRx/Act/Tx":"_evpn"}}',
-            "datatype": "state",
-            "key": "index",
-        }
-        resp = self.get(
-            paths=[path_spec.get("path", "")], datatype=path_spec["datatype"]
-        )
-        augment_resp(resp)
-        res = jmespath.search(path_spec["jmespath"], resp[0])
-        return {"bgp_peers": res}
 
     def get_lldp_sum(self, interface: Optional[str] = "*") -> Dict[str, Any]:
         path_spec = {
@@ -722,7 +342,6 @@ class SrLinux(NetworkInstanceMixin):
 
         res = jmespath.search(path_spec["jmespath"], resp[0])
         return {"ip_rib": res}
-
 
     def get(
         self,
